+++ packages/core-js/modules/esnext.uint8-array.from-hex.js
@@ -1,28 +1,13 @@
 'use strict';
 var $ = require('../internals/export');
 var global = require('../internals/global');
-var uncurryThis = require('../internals/function-uncurry-this');
 var aString = require('../internals/a-string');
-
-var Uint8Array = global.Uint8Array;
-var SyntaxError = global.SyntaxError;
-var parseInt = global.parseInt;
-var NOT_HEX = /[^\da-f]/i;
-var exec = uncurryThis(NOT_HEX.exec);
-var stringSlice = uncurryThis(''.slice);
+var $fromHex = require('../internals/uint8-from-hex');
 
 // `Uint8Array.fromHex` method
 // https://github.com/tc39/proposal-arraybuffer-base64
-if (Uint8Array) $({ target: 'Uint8Array', stat: true }, {
+if (global.Uint8Array) $({ target: 'Uint8Array', stat: true }, {
   fromHex: function fromHex(string) {
-    aString(string);
-    var stringLength = string.length;
-    if (stringLength % 2) throw new SyntaxError('String should have an even number of characters');
-    if (exec(NOT_HEX, string)) throw new SyntaxError('String should only contain hex characters');
-    var result = new Uint8Array(stringLength / 2);
-    for (var i = 0; i < stringLength; i += 2) {
-      result[i / 2] = parseInt(stringSlice(string, i, i + 2), 16);
-    }
-    return result;
+    return $fromHex(aString(string)).bytes;
   }
 });